Refactor raw type method to safe type. Book globalscore working.

+++ src/main/java/bookers/bookkeeper/booklist/BookEntry.java
@@ -9,6 +9,7 @@ import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import lombok.Data;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.sql.Date;
 
 @Entity
@@ -34,7 +35,7 @@ public class BookEntry {
     private User user;
 
     @Column(name = "score")
-    private Float userScore;
+    private Double userScore;
 
     @Column(name = "pages_read")
     private Integer pagesRead;
